feat: node-wide settlement engine url map
this adds a node-wide configuration method to set a mapping from
asset_code to settlement_engine_url. if an account does not have a
settlement_engine_url set but there is one configured for the account's
asset_code, the globally configured one will be set as the account's
settlement_engine_url.